+++ Customer.Inquiry.BusinessLogic/Interface/IBusinessLogicBase.cs
@@ -7,9 +7,9 @@ namespace Customer.Inquiry.BusinessLogic.Interface
     public interface IBusinessLogicBase<T> where T : class, IDomainBase
     {
         Task<IList<T>> GetList();
-        Task<T> Get(int internalId);
+        Task<T> Get(long internalId);
         Task<T> Update(T item);
         Task<T> Save(T item);
-        void Delete(int internalId);
+        void Delete(long internalId);
     }
 }
