+++ Customer.Inquiry.Repository/Interface/IRepositoryBase.cs
@@ -7,9 +7,9 @@ namespace Customer.Inquiry.Repository.Interface
     public interface IRepositoryBase<T>
     {
         Task<IList<T>> GetList();
-        Task<T> Get(int internalId);
+        Task<T> Get(long internalId);
         Task<T> Update(T item);
         Task<T> Save(T item);
-        void Delete(int internalId);
+        void Delete(long internalId);
     }
 }
